issue - #876 - added auditConfig check - refactored tests for audit - adjusted tests for audit - removed unused handler.yml
