--- audit/src/test/resources/handler.yml
@@ -1,11 +1,0 @@
----
-enabled: true
-
-# Configuration for the LightHttpHandler. The handler is the base class  for all middleware, server and health handlers
-# set the Status Object in the AUDIT_INFO, for auditing purposes
-# default, if not set:false
-auditOnError: true
-
-# set the StackTrace in the AUDIT_INFO, for auditing purposes
-# default, if not set:false
-auditStackTrace: true
